Completely avoid usage of stubs in app layer
While trying to add stuff that's not already public in
SoftApConfiguration in Android SDK, although Gradle can build the
project, Android Studio cries about property/method not found as it
resolves Android SDK's sources first and foremost which leads to one
hell of an annoying coding session. For this reason, I added hidden
api refine plugin provided by the same devs who made Shizuku which
forced me to do unsafe casts between data types (actual obj to stub).
This seeped in all the way to the app code which breaks all the
reasons for separating the stubs, the API and the UI layer, the
entire impl becomes messy.

Now, the API layer has wrapper classes wherever needed and the app
subproject is not linked with system-api-stubs at all.